backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy<2.0.0
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0